Initial commit for EJP-RD branch
This branch is being used for extensions to the FHIR generator that
modify the data model to make it useful for the EJP-RD project.
This basically means adding things that are relevant to rare
diseases. I have been careful to ensure that nothing in the original
GBA data model is removed. Additionally, this feature is not
turned on by default. In order to use it, the user must explicitly
provide a "-e" option on the command line. Without this option,
the code works in exactly the same way as the code on the
master branch.

Roughly, what has been added are:

* A new FHIR-Resource, FamilyMemberHistory, that records information
  about relatives of a patient.
* A new code for Observation: Orphanet.
* A new code for Observation: HP diagnosis.
* An extension to Observation to hold a Genotype string.
* Modifications to Patient, primarily so that it also contains
  name and address information for the patient.